+++ api/src/pdc/proxy/helpers/castContext.ts
@@ -1,5 +1,5 @@
 import { ContextType } from "@/ilos/common/index.ts";
-import { UserInterface } from "@/shared/user/common/interfaces/UserInterface.ts";
+import { UserInterface } from "../../services/user/contracts/common/interfaces/UserInterface.ts";
 
 export function castContext(
   user?: Partial<UserInterface>,
