+++ api/src/pdc/proxy/helpers/createRPCPayload.ts
@@ -1,5 +1,5 @@
 import { ParamsType, RPCSingleCallType } from "@/ilos/common/index.ts";
-import { UserInterface } from "@/shared/user/common/interfaces/UserInterface.ts";
+import { UserInterface } from "../../services/user/contracts/common/interfaces/UserInterface.ts";
 import { injectContext } from "./injectContext.ts";
 
 export function createRPCPayload(
