+++ api/src/pdc/services/application/ApplicationServiceProvider.ts
@@ -3,15 +3,15 @@ import { ServiceProvider as AbstractServiceProvider } from "@/ilos/core/index.ts
 import { defaultMiddlewareBindings } from "@/pdc/providers/middleware/index.ts";
 import { ValidatorExtension, ValidatorMiddleware } from "@/pdc/providers/validator/index.ts";
 
-import { binding as createBinding } from "@/shared/application/create.schema.ts";
-import { binding as findBinding } from "@/shared/application/find.schema.ts";
-import { binding as listBinding } from "@/shared/application/list.schema.ts";
-import { binding as revokeBinding } from "@/shared/application/revoke.schema.ts";
 import { CreateApplicationAction } from "./actions/CreateApplicationAction.ts";
 import { FindApplicationAction } from "./actions/FindApplicationAction.ts";
 import { ListApplicationAction } from "./actions/ListApplicationAction.ts";
 import { RevokeApplicationAction } from "./actions/RevokeApplicationAction.ts";
 import { config } from "./config/index.ts";
+import { binding as createBinding } from "./contracts/create.schema.ts";
+import { binding as findBinding } from "./contracts/find.schema.ts";
+import { binding as listBinding } from "./contracts/list.schema.ts";
+import { binding as revokeBinding } from "./contracts/revoke.schema.ts";
 import { ApplicationPgRepositoryProvider } from "./providers/ApplicationPgRepositoryProvider.ts";
 
 @serviceProvider({
