+++ api/src/pdc/services/application/contracts/common/interfaces/ApplicationCreateResultInterface.ts
