+++ api/src/pdc/services/application/interfaces/ApplicationRepositoryProviderInterface.ts
@@ -1,8 +1,8 @@
-import { ApplicationInterface } from "@/shared/application/common/interfaces/ApplicationInterface.ts";
-import { RepositoryInterface as ListInterface } from "@/shared/application/list.contract.ts";
-import { RepositoryInterface as FindInterface } from "@/shared/application/find.contract.ts";
-import { RepositoryInterface as CreateInterface } from "@/shared/application/create.contract.ts";
-import { RepositoryInterface as RevokeInterface } from "@/shared/application/revoke.contract.ts";
+import { ApplicationInterface } from "../contracts/common/interfaces/ApplicationInterface.ts";
+import { RepositoryInterface as CreateInterface } from "../contracts/create.contract.ts";
+import { RepositoryInterface as FindInterface } from "../contracts/find.contract.ts";
+import { RepositoryInterface as ListInterface } from "../contracts/list.contract.ts";
+import { RepositoryInterface as RevokeInterface } from "../contracts/revoke.contract.ts";
 
 export interface ApplicationRepositoryProviderInterface {
   list(data: ListInterface): Promise<ApplicationInterface[]>;
@@ -11,8 +11,7 @@ export interface ApplicationRepositoryProviderInterface {
   revoke(data: RevokeInterface): Promise<void>;
 }
 
-export abstract class ApplicationRepositoryProviderInterfaceResolver
-  implements ApplicationRepositoryProviderInterface {
+export abstract class ApplicationRepositoryProviderInterfaceResolver implements ApplicationRepositoryProviderInterface {
   async list(data: ListInterface): Promise<ApplicationInterface[]> {
     throw new Error("Method not implemented.");
   }
